+++ cmd/manage/main.go
@@ -3,11 +3,11 @@ package main
 import (
 	"os"
 
-	"github.com/OpenSlides/openslides-manage-service/pkg/client"
+	"github.com/OpenSlides/openslides-manage-service/pkg/manage"
 )
 
 func main() {
-	if err := client.Execute(); err != nil {
+	if err := manage.RunClient(); err != nil {
 		os.Exit(1)
 	}
 }
